backend/pyproject.toml
[tool.poetry]
name = "coursework-2024-backend"
version = "0.2.0"
description = ""
authors = ["Arseniy Krupchik <kruchik.a04@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.3"
fastapi = "^0.110.2"
SQLAlchemy = "^2.0.25"
alembic = "^1.13.1"
uvicorn = "^0.20.0"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
mimesis = "^13.1.0"
pytest = "^7.4.4"
mypy = "^1.11.1"
pylint = "^3.2.6"

[tool.mypy]
strict = true
python-version = "3.12"
exclude = ["alembic/"]

[tool.ruff]
line-length = 79
indent-width = 4
target-version = "py312"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
